:construction: recursive mooreN fun
The handling of the exceptions on mooreN is via a recursion.

The strategy is the following:

Try to grow the N-1 grid into the N grid by growing 1 step
on each the 4 sides (up, down, left, right).

The grew sides could come from a source global GRID
or be a default value
(to default when out of borders on GRID).
The grew sides have a size of N-1.

Then check the corners to finish the bulding of the N size grid.

The base case is the when N == 0, on that case
the output is the single cell at the selected position.

This algorithm asumes that the target postion exists.
It should be easy to refactor into an iterative form,
and maybe is easier to understand.