:recycle: Remove get_path
Until function get_path was removed.

Getting the current global path to the project only once makes more sense.
Any procedure that needs to calculate a path can do it relative to the global path.

The update_gallery script was executed, generating the new renders for the envs.
Those renders will be shown on the README front page.