:hammer: versionate logic
Upgrading major or minor version numbers
sets to zero all version number below.

For example updating the major version number from 1 to 2
will set to 0 the minor and patch numbers.

The logic of downgrading versions is unchanged.

The precommit hook of checking test files names was dropped.
To inconvenient to name all the test directories. If the hook is required
it could be done as a command line one-liner.