:adhesive_bandage: move operator fix
Minor fix on move operator.

The attribute grid_dependant was changed to False.

If assuming grids with variable size as inputs,
then the grid has an effect on the output,
however this is not the intended use.

On the gym API the observation size is constant,
thus the grid size is constant.