:recycle: separate registration
Registration logic was separated into a file.
This cleans up the library init file.

The registration data was gathered into a dictionary.
This separates the data from the actual register execution.
The register is now a loop that request the data.

The gym make strings are automatically generated and
place into a list for user & developer convenience.