:hammer: clean & parallel gallery
Updated Makefile to clean temporal files.
gymca uses the tag TMP at the start of a file to signal it as a temporal.

update_gallery strategy to not overwrite files it is not guaranteed and depends on randomness. However the chance of overwriting a file is tiny.
This is on purpose, as it is an easy hack to parallelize the script.