:arrow_up: remove pip installs & set to 3.9
Removed pip installs on the conda environment yaml file.
This makes the environment more stable as conda dislikes
installs from pip.

Added make rule to remove the gymca environment.

The conda environment was set-up to python 3.9,
3.10 still to new and needs further testing.

I am considering to support gymca library from python 3.7,
as gym currently does, I think it is just a matter of type annotation edits.

Tests are failing due to upgrade to gym 0.21.0
the spaces method _contain_ it is now stricter,
so comparing floats32 with floats64 throws an error
instead of a warning like the previous version.