:recycle: blldz single def rows & cols
Refactor to read the nrows and ncols from a single source.

Accounting to software engineering is better to delay taking
decisions if they can cleanly be done on upper layers.
Something like keeping your options open.

The nrows and ncols parameters will be singly decided on the registration code.

This refactor is pending for helicopter env.