:adhesive_bandage: env to 3.9
Python version must be on environment.yaml
instead of the conda command, for whatever reason this
`conda env create --file python=3.9`
does not work but also does not raises an error.

It was fixed by adding this line
`- python=3.9`
into the yaml file
and removing the version on the command.