:recycle: TYPE_BOX global
The float type of space Box is extracted at run time,
and made it available as a global variable.

The used type is the current default of gym of _float32_

The idea is to abstract away the explicit typing
of actions and observations while being consistent with
space.contains method.